% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrib_data.R
\docType{data}
\name{attrib_ts_data}
\alias{attrib_ts_data}
\title{Attribution and intellectual rights information for MacroSheds time-series and watershed attribute data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 414 rows and 27 columns.
}
\usage{
attrib_ts_data
}
\description{
Also includes contact info and more, organized by primary source data product
}
\author{
Mike Vlah \email{vlahm13@gmail.com}
}
\keyword{data}
