% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ms_site_data.R
\name{download_ms_site_data}
\alias{download_ms_site_data}
\title{Download macrosheds site data table}
\usage{
download_ms_site_data()
}
\value{
returns a \code{data.frame} of all sites in the macrosheds systems 
   with their coresponding network, domain, latitude, longitude, area, and 
   aditional information
}
\description{
Download the macrosheds site data table with information on networks, domains,
 location of stream gauges, and more
}
\examples{
macrosheds_sites <- download_ms_site_data()
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
