% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_product.R
\name{load_product}
\alias{load_product}
\title{Load in macrosheds data products}
\usage{
load_product(
  macrosheds_root,
  prodname,
  filter_vars,
  domains,
  sort_result = FALSE
)
}
\arguments{
\item{macrosheds_root}{character. The path to the macrosheds dataset's parent
directory, e.g. '~/stuff/macrosheds_dataset_v0.3'}

\item{prodname}{character. read and combine all files associated with this prodname
across all networks and domains. Available prodnames are:
discharge, stream_chemistry, stream_flux_inst, precipitation,
precip_chemistry, precip_flux_inst.}

\item{filter_vars}{character vector. for products like stream_chemistry that include
multiple variables, this filters to just the ones specified (ignores
variable prefixes). To see a catalog of variables, visit macrosheds.org}

\item{domains}{character vector. Domains to load, optional}

\item{sort_result}{logical. If TRUE, output will be sorted by site_code, var,}
}
\value{
returns a \code{tibble} containing all the all data belonging to the 
   selected data product and variables in the \code{macrosheds_root} directory
}
\description{
Load an entire macrosheds product (e.g stream_chemistry, discharge, etc.) from a 
downloaded macrosheds dataset see \code{download_ms_code_data()} for data download
}
\examples{
macrosheds_data <- load_product(macrosheds_root = 'data/macrosheds_v1', 
                                prodname = 'stream_chemistry', 
                                sort_result = FALSE, 
                                filter_vars = 'NO3_N')
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
