% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spatial_products.R
\name{load_spatial_products}
\alias{load_spatial_products}
\title{Load in macrosheds spatial data products}
\usage{
load_spatial_products(macrosheds_root, spatial_product, domains, site_codes)
}
\arguments{
\item{macrosheds_root}{character. The path to the macrosheds dataset's parent
directory, e.g. '~/stuff/macrosheds_dataset_v0.3'}

\item{spatial_product}{character. read and combine all files associated with 
this spatial prodname across all networks and domains. Available spatial_product 
are: ws_boundary, precip_gauge_locations, or stream_gauge_locations}

\item{domains}{character vector. Domains to load, optional}

\item{site_codes}{character vector. site_codes to load, optional}
}
\value{
returns a \code{sf} object containing all the all data belonging to the 
   selected spatial product in the \code{macrosheds_root} directory
}
\description{
Load an entire macrosheds spatial product (ws_boundary, precip_gauge_locations, 
or stream_gauge_locations) from a downloaded macrosheds dataset see 
\code{download_ms_code_data()} for data download
}
\examples{
macrosheds_data <- load_spatial_products(macrosheds_root = 'data/macrosheds_v1', 
                                         spatial_product = 'ws_boundary'
                                         domains = 'hbef')
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
