% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_calc_flux.R
\name{ms_calc_flux}
\alias{ms_calc_flux}
\title{Calculate daily solute fluxes}
\usage{
ms_calc_flux(chemistry, q)
}
\arguments{
\item{chemistry}{A \code{data.frame} of precipitation or
stream chemistry data in MacroSheds format (see details) and in units of mg/L.}

\item{q}{A \code{data.frame} in MacroSheds format (see details) containing
stream discharge in L/s or precipitation in mm.}
}
\value{
returns a \code{tibble} in MacroSheds format (see details) of stream or
precipitation chemical flux for every dayduring which discharge/precipitation
and chemistry are reported. Output units are kg/ha/d.
}
\description{
Determines solute fluxes from interpolated daily Q (discharge
or precipitation) and corresponding chemistry data.
For cumulative solute loads computed by various methods, see \code{\link[=ms_calc_load]{ms_calc_load()}}.
}
\details{
#' MacroSheds format (only date, site_code, var, and val are required in inputs to this function):\tabular{ll}{
   column \tab definition \cr
   date \tab Date in YYYY-mm-dd \cr
   site_code \tab A unique identifier for each MacroSheds site, identical to primary source site code where possible. See \code{\link[=ms_load_sites]{ms_load_sites()}}. \cr
   grab_sample \tab Boolean integer indicating whether the observation was obtained via grab sample or installed sensor. 1 = TRUE (grab sample), 0 = FALSE (installed sensor). \cr
   var \tab Variable code. See \code{\link[=ms_load_variables]{ms_load_variables()}}. \cr
   val \tab Data value. See \code{\link[=ms_load_variables]{ms_load_variables()}} for units. \cr
   ms_status \tab Boolean integer. 0 = clean value. 1 = questionable value. See "Technical Validation" section of \href{https://aslopubs.onlinelibrary.wiley.com/doi/full/10.1002/lol2.10325}{the MacroSheds data paper} for details. \cr
   ms_interp \tab Boolean integer. 0 = measured or imputed by primary source. 1 = interpolated by MacroSheds. See "Temporal Imputation and Aggregation" section of \href{https://aslopubs.onlinelibrary.wiley.com/doi/full/10.1002/lol2.10325}{the MacroSheds data paper} for details. \cr
   val_err \tab The combined standard uncertainty associated with the corresponding data point, if estimable. See "Detection Limits and Propagation of Uncertainty" section of \href{https://aslopubs.onlinelibrary.wiley.com/doi/full/10.1002/lol2.10325}{the MacroSheds data paper} for details. \cr
}


This is the format returned by ms_load_product for core time-series data.

Our Solute fluxes are calculated by multiplying solute concentration by daily water volume.
Stream and precipitation flux are computed differently. For stream flux (stream chemistry * discharge),
flux is: kg/ha/d = mg/L \* L/s \* 86400 / 1e6 / ws_area.
Precipitation flux (precip chemistry \* precip depth over the watershed)
is calculated as: kg/ha/d = mg/L \* mm/d / 100.
You can convert between kg/ha/d and kg/d using \code{\link[=ms_scale_flux_by_area]{ms_scale_flux_by_area()}} and
\code{\link[=ms_undo_scale_flux_by_area]{ms_undo_scale_flux_by_area()}}.
}
\examples{
ms_root = 'data/macrosheds'
ms_download_core_data(macrosheds_root = ms_root,
                      domains = 'hbef')

chemistry <- ms_load_product(macrosheds_root = ms_root,
                             prodname = 'stream_chemistry',
                             site_codes = c('w1', 'w3'),
                             filter_vars = c('NO3_N', 'Cl'))

q <- ms_load_product(macrosheds_root = ms_root,
                     prodname = 'discharge',
                     site_codes = c('w1', 'w3'))

flux <- ms_calc_flux(chemistry = chemistry,
                     q = q)
}
\seealso{
\code{\link[=ms_calc_load]{ms_calc_load()}}, \code{\link[=ms_scale_flux_by_area]{ms_scale_flux_by_area()}}, \code{\link[=ms_undo_scale_flux_by_area]{ms_undo_scale_flux_by_area()}}
}
\author{
Spencer Rhea

Mike Vlah, \email{vlahm13@gmail.com}

Wes Slaughter
}
