% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_calc_inst_flux.R
\name{ms_calc_inst_flux}
\alias{ms_calc_inst_flux}
\title{Calculates chemical fluxes}
\usage{
ms_calc_inst_flux(chemistry, q, q_type, verbose = TRUE)
}
\arguments{
\item{chemistry}{\code{data.frame}. A \code{data.frame} of precipitation or
stream chemistry data in macrosheds format and in units of mg/l.}

\item{q}{\code{data.frame}. A \code{data.frame} of precipitation or stream
discharge in macrosheds format and in units of mm or L/s respectively.}

\item{verbose}{logical. Default true, should function print information to
console.}

\item{data_type}{character. Either precipitation or discharge.}
}
\value{
returns a \code{tibble} of stream instantaneous flux for every timestep
where discharge/precipitation and chemistry are reported in units of kg/ha/timestep.
}
\description{
Calculates stream discharge and precipitation fluxes of chemicals from Q (discharge
or precipitation) and chemistry data
}
\details{
Chemical flux is calculated by multiplying chemical concentration by flow
of water (flux = concentration * flow). The output units depend on the time
interval the input data is collected at. If data is collected at 15 minutes
the out put will be in kg/ha/15 minutes and if collected daily units will
be in kg/day.

Before imputing data into \code{ms_calc_inst_flux()}, insure all both q and
chemistry data is in the same time step. See \code{ms_synchronize_timestep} to
take data of different intervals and interpolate or summarize data to different
sampling intervals. Also insure chemistry units are in mg/l, see \code{ms_conversions}
for unit conversions.
}
\examples{

ms_download_core_data(macrosheds_root = 'data/ms_test',
                      domains = 'hbef')
chemistry <- macrosheds::ms_load_product('data/ms_test/',
                                         'stream_chemistry',
                                         site_codes = c('w1', 'w3', 'w6'),
                                         filter_vars = c('NO3_N', 'Cl', 'Na'),
                                         warn = F)

q <- macrosheds::ms_load_product('data/ms_test/',
                                 'discharge',
                                 site_codes = c('w1', 'w3', 'w6'),
                                 warn = F)

flux <- ms_calc_inst_flux(chemistry = chemistry, q = q, q_type = 'discharge')
}
\author{
Spencer Rhea, spencerrhea41@gmail.com

Mike Vlah

Wes Slaughter
}
