% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_catalog.R
\name{ms_catalog}
\alias{ms_catalog}
\title{Download the MacroSheds site-variable catalog}
\usage{
ms_catalog(fp = NULL)
}
\arguments{
\item{fp}{character. provide a valid filepath to download the catalog to.}
}
\value{
returns a \code{data.frame} of the variables in
the MacroSheds dataset, the sites that record them,
number of observations, and more.
}
\description{
downloads the MacroSheds catalog, which documents
important metadata about what variables are recorded at what sites,
and the number of observations, first and last record, and other
important metadata for all site-variable pairs.
}
\examples{
site_vars_catalog <- ms_catalog()
}
\seealso{
\code{\link[=ms_download_variables]{ms_download_variables()}}, \code{\link[=ms_download_site_data]{ms_download_site_data()}}
}
\author{
Spencer Rhea, \email{spencerrhea41@@gmail.com}

Mike Vlah

Wes Slaughter
}
