% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_conversions.R
\name{ms_conversions}
\alias{ms_conversions}
\title{Convert macrosheds concentration units}
\usage{
ms_conversions(
  d,
  convert_units_from = "mg/l",
  convert_units_to,
  convert_molecules
)
}
\arguments{
\item{d}{\code{date.frame}. A macrosheds \code{dataframe} of concentration data 
downloaded using \code{download_ms_core_data()}.}

\item{convert_units_from}{a named character vector or character.
Names are variable names without their sample-regimen prefixes 
(e.g. 'DIC', not 'GN_DIC'), and values are the units of those variables. 
Omit variables that don't need to be converted. Or a single unit if all 
variables are in the same unit.}

\item{convert_units_to}{a named character vector or character. Names are 
variable names without their sample-regimen prefixes (e.g. 'DIC', not 'GN_DIC'),
and values are the units those variables should be converted to.
Omit variables that don't need to be converted. Or a single unit if all 
variables are in the same unit.}

\item{convert_molecules}{character vector. A vector of variables that will be 
converted from or to the atomic mass of a molecule's main constituent 
(NO3_N to NO3 or NO3 to NO3_N), see details.}
}
\value{
returns a \code{tibble} containing concentration data converted to new units
}
\description{
Convert macrosheds data from it's native format into units of g/l, mol/l, and 
eq/l. Along with metric base unit conversions and conversions from molecules 
in the form of NO3 as N (NO3_N) to NO3 (NO3) and vice versa
}
\examples{
converted_data <- ms_conversions(d = d,
                                 convert_units_from = 'mg/l',
                                 convert_units_to = 'ug/l')
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
