% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_download_core_data.R
\name{ms_download_core_data}
\alias{ms_download_core_data}
\title{Download macrosheds core datasets}
\usage{
ms_download_core_data(macrosheds_root, networks, domains, quiet = FALSE)
}
\arguments{
\item{macrosheds_root}{character.  Directory where macrosheds data files will be downloaded to}

\item{networks}{character vector. macrosheds networks that will be downloaded. 
Either a single network, vector of networks, or 'all'. see \code{download_ms_site_data()} 
for networks available for download.}

\item{domains}{character vector. macrosheds domains that will be downloaded. 
Either a single domain, vector of domains, or 'all'. see \code{download_ms_site_data()} 
for domains available for download.}

\item{quiet}{logical. If TRUE, no download messages will be printed in console}
}
\value{
Downloads all core data for selected domains to the designated 
   directory listed in \code{macrosheds_root}. Data follows the structure: domain/product/site_code.feather. 
   Where domain is the domain selected, product is one or more of the data 
   product available at a domain (eg. stream_chemistry__ms001, discharge__ms002), 
   and site_code is the code for an individual site. See \code{download_ms_site_data()} 
   for more information on sites.
}
\description{
Download the macrosheds core datasets of stream chemistry, stream 
discharge, stream chemical flux, precipitation, precipitation chemistry, and
precipitation chemical flux. Not all products are available at all sites but 
all products available for the selected domains will be downloaded.
}
\details{
Only \code{networks} or \code{domains} must be supplied. If 'all' is 
   supplied to either argument, all domains will be downloaded regardless of 
   what is supplied to the other argument. Downloading all macrosheds core data
   can take up to ~5 GB, so beware of downloading too much data.
}
\examples{
dir.create('data/macrosheds')
download_ms_core_data(macrosheds_root = 'data/macrosheds',
                      domains = c('niwot', 'hjandrews'))
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
