% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_download_site_data.R
\name{ms_download_site_data}
\alias{ms_download_site_data}
\title{Download macrosheds site data table}
\usage{
ms_download_site_data(fp = NULL)
}
\value{
returns a \code{data.frame} of all sites in the macrosheds systems
with their corresponding network, domain, latitude, longitude, area, and
additional information.
}
\description{
Download the macrosheds site data table with information on networks, domains,
location of stream gauges, and more.
}
\examples{
<<<<<<< HEAD
macrosheds_sites <- download_ms_site_data()
=======
macrosheds_sites <- ms_download_site_data()
}
\seealso{
\code{\link[=ms_download_variables]{ms_download_variables()}}, \code{\link[=ms_download_core_data]{ms_download_core_data()}}
>>>>>>> 43d097da872114e008e3cf37e874dddc04a25dc8
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
