% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_download_variables.R
\name{ms_download_variables}
\alias{ms_download_variables}
\title{Download macrosheds variable metadata (deprecated)}
\usage{
ms_download_variables()
}
\description{
This function has been retired, and replaced with \code{ms_load_variables}. Please use this instead.
}
\seealso{
\code{\link[=ms_load_variables]{ms_load_variables()}}
}
\author{
Wes Slaughter, \email{wslaughter@berkeley.edu}

Mike Vlah, \email{vlahm13@gmail.com}

Spencer Rhea
}
