% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_download_ws_attr_summary.R
\name{ms_download_ws_attr_summary}
\alias{ms_download_ws_attr_summary}
\title{Download MacroSheds watershed attribute summary data tables}
\usage{
ms_download_ws_attr_summary(macrosheds_root)
}
\value{
returns a \code{data.frame} of all sites in the MacroSheds systems
with their corresponding network, domain, latitude, longitude, area, and
additional information.
}
\description{
Download the MacroSheds site data table with information on networks, domains,
location of stream gauges, and more. (source .Rdata file downloaded with the MacroSheds package)
}
\examples{
ms_site_data <- ms_download_ws_attr_summary()
}
\seealso{
\code{\link[=ms_download_site_data]{ms_download_site_data()}}, \code{\link[=ms_download_core_data]{ms_download_core_data()}}
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
