% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_drop_var_prefix.R
\name{ms_drop_var_prefix}
\alias{ms_drop_var_prefix}
\title{Drop MacroSheds variable prefixes (DEPRECATED)}
\usage{
ms_drop_var_prefix(x)
}
\arguments{
\item{x}{character. MacroSheds variable code with a prefix.}
}
\value{
Returns a vector of MacroSheds variable codes without their two character prefixes.
If prefixes are missing, returns \code{x} with an warning.
}
\description{
Remove the two-letter prefixes from MacroSheds variable codes. Retained for
legacy use, but will be removed in a later version. See details.
}
\details{
In version 1 of the MacroSheds dataset, variable codes all included a two-letter prefix to denote the
sampling regimen. The first letter is either I for "installed"
or G for "grab," denoting if this variable was recorded by a continuously
monitoring piece of equipment, or manually by a human.
The second position is either S for "sensor" or N for "non-sensor".
Non-sensor samples are generally measured analytically, in a laboratory.
In version 2 and later, this system has been simplified and replaced by a single,
logical column called "grab_sample".
}
\examples{
macrosheds_vars <- c('GN_NO3_N', 'GN_Mg', 'IS_discharge')
macrosheds_vars <- drop_var_prefix(x = macrosheds_vars)
}
\seealso{
\code{\link[=ms_extract_var_prefix]{ms_extract_var_prefix()}}
}
\author{
Spencer Rhea

Mike Vlah, \email{vlahm13@gmail.com}

Wes Slaughter
}
