% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_extract_var_prefix.R
\name{ms_extract_var_prefix}
\alias{ms_extract_var_prefix}
\title{Extract macrosheds variable prefixes}
\usage{
ms_extract_var_prefix(x)
}
\arguments{
\item{x}{character string. Macrosheds variable with a prefix.}
}
\value{
returns the two character prefix on macrosheds variables.
}
\description{
Extract prefixes on macrosheds variables.
}
\details{
Macrosheds variables all have a two letter prefix on the front of the 
   variable to denote the sampling type. The first space is either a I for 'installed'
   or a G for 'grab.' Denoting if this variable records continuously or is a grab 
   sample. The second position is either a S for 'sensor' or N for 'not a sensor'.
   Non-sensor variables are generally analyzed in a laboratory.
}
\examples{
macrosheds_vars <- c('GN_NO3_N', 'GN_Mg', 'IS_discharge')
macrosheds_prefixs <- extract_var_prefix(x = macrosheds_vars)
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
