% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_identify_usgs_gauges.R
\name{ms_identify_usgs_gauges}
\alias{ms_identify_usgs_gauges}
\title{Identify USGS gauges near a point}
\usage{
ms_identify_usgs_gauges(sites, lat = NULL, long = NULL, radius = 500)
}
\arguments{
\item{sites}{Either a \code{sf} object of sites of interest or a \code{data.frame}
containing latitude and longitude information.}

\item{lat}{Character. If sites is not a \code{sf} object, the name of the column
containing the latitude locations in WGS84. Default NULL.}

\item{long}{Character. If sites is not a \code{sf} object, the name of the column
containing the longitude locations in WGS84. Default NULL.}

\item{radius}{Numeric. The radius in meters that a USGS gauge will be search
for from a point in the sites file. Default 500.}
}
\value{
returns a \code{data.frame} with a the column usgs_site.
}
\description{
Identify if there any USGS gauges in a radius of a latitude and longitude location
}
\details{
The function will search for a USGS gauge within the radius supplied
for every site in the sites file. The dataRetrieval package is used to locate USGS gauges.
If no USGS gauge is found, then the usgs_site column will be NA for that site.
}
\examples{
# With lat long 
sites <- tibble(site_code = c('MA_AE03', 'MI_KR01'),
                Latitude = c(42.04856, 42.28575),
                Longitude = c(-72.45403, -85.51467))
ms_identify_usgs_gauges(sites = sites, lat = 'Latitude', long = 'Longitude')
# With sf
site_sf <- sites \%>\%
    sf::st_as_sf(coords = c('long', 'lat'), crs = 4326) \%>\%
    sf::st_buffer(radius)
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
