% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_load_sites.R
\name{ms_load_sites}
\alias{ms_load_sites}
\title{Load MacroSheds site data table}
\usage{
ms_load_sites()
}
\value{
returns a \code{data.frame} of all sites in the MacroSheds dataset
with their corresponding network, domain, latitude, longitude, area, and
additional information.
}
\description{
Load the MacroSheds site data table with information on networks, domains,
location of stream gauges, and more.
}
\details{
Site metadata are included as an .RData file with the macrosheds package, and
are loaded into the macrosheds namespace on library(macrosheds). This
function simply retrieves a data.frame from the macrosheds namespace and makes it available
in the global environment.
}
\examples{
site_data <- ms_load_sites()
}
\seealso{
\code{\link[=ms_load_variables]{ms_load_variables()}}, \code{\link[=ms_load_spatial_product]{ms_load_spatial_product()}}, \code{\link[=ms_load_product]{ms_load_product()}}
}
\author{
Wes Slaughter, \email{weston.slaughter@duke.edu}

Mike Vlah

Spencer Rhea
}
