% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_load_variables.R
\name{ms_load_variables}
\alias{ms_load_variables}
\title{load MacroSheds variables data table}
\usage{
ms_load_variables(var_set = "timeseries")
}
\arguments{
\item{expanded}{logical. If TRUE, ms_load_variables will load in the complete variable sheet which includes dozens of
more obscure molecules.}
}
\value{
returns a \code{data.frame} of variables, codes,
and units in MacroSheds format
}
\description{
Load a data.frame of MacroSheds variables with information on variable codes,
units, and more.
}
\details{
Variable metadata are included as an .RData file with the macrosheds package, and
are loaded into the macrosheds namespace on library(macrosheds). This
function simply retrieves a data.frame from the macrosheds namespace and makes it available
in the global environment.
}
\examples{
vars <- ms_load_variables()
}
\seealso{
\code{\link[=ms_load_sites]{ms_load_sites()}}, \code{\link[=ms_load_spatial_product]{ms_load_spatial_product()}}, \code{\link[=ms_load_product]{ms_load_product()}}
}
\author{
Wes Slaughter, \email{weston.slaughter@duke.edu}

Mike Vlah

Spencer Rhea
}
