% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_precip_interp.R
\name{ms_precip_interp}
\alias{ms_precip_interp}
\title{Interpolate precipitation gauge data to whole basins}
\usage{
ms_precip_interp(
  precip,
  ws_boundary,
  precip_gauge,
  pchem,
  parallel = TRUE,
  maxcores = Inf,
  out_path,
  elevation_agnostic = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{precip}{\code{data.frame} or path to a folder containing \code{data.frame}
files of precipitation gauge data in macrosheds format.}

\item{ws_boundary}{\code{sf} object or path to a folder containing \code{sf}
files of watershed boundaries in macrosheds format.}

\item{precip_gauge}{\code{sf} object or path to a folder containing \code{sf}
files of precipitation gauge locations in macrosheds format.}

\item{pchem}{\code{data.frame} or path to a folder containing \code{data.frame}
files of precipitation chemistry data in macrosheds format.}

\item{parallel}{logical. Default TRUE, should interpolation run in parallel.}

\item{maxcores}{numeric. Default Inf, the maximum number of cores used in parallel processing,
ignored if parallel is FALSE.}

\item{out_path}{character. Directory to where precipitation data will be saved.}

\item{elevation_agnostic}{Logical. Should elevation data be used in interpolating
precipitation data. See details for more information}

\item{verbose}{Logical. Should information be printed to console.}
}
\value{
Saves precipitation, precipitation chemistry, and flux (if both precip
and pchem are supplied) to the out_path.
}
\description{
Interpolate daily precipitations volume, chemistry, and precipitation chemistry flux
to basins in a study area.
}
\details{
ms_precip_interp, using inverse weighted interpolation, calculates mean
watershed precipitation. ms_precip_interp will interpolate only precipitation volume if no
pchem is supplied. The function will also only interpolate precipitation chemistry
if no precip data is supplied. If both precip and pchem are supplied, then
precipitation volume, chemical concentrations, and precipitation chemical fluxes
will also be calculated.

When elevation_agnostic is TRUE, a liner regression between precipitation gauge
elevation and precipitation volume will be developed for every time step. Then
precipitation will be interpolated across the basin using inverse weighted
interpolation with gauge location and the linear regression will be used to
fill raster cells based on their elevation. Finally a mean between the inverse
weighted interpolation and elevation relationship will be used for each time step.
When FALSE, only inverse weighted interpolation is used.
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
