% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_read_csv.R
\name{ms_read_raw_csv}
\alias{ms_read_raw_csv}
\title{read a csv into macrosheds format}
\usage{
ms_read_raw_csv(
  filepath,
  preprocessed_tibble,
  datetime_cols,
  datetime_tz,
  optionalize_nontoken_characters = ":",
  site_code_col,
  alt_site_code,
  data_cols,
  data_col_pattern,
  alt_datacol_pattern,
  is_sensor,
  set_to_NA,
  var_flagcol_pattern,
  alt_varflagcol_pattern,
  summary_flagcols,
  sampling_type = NULL
)
}
\arguments{
\item{datetime_cols}{}

\item{datetime_tz}{}

\item{optionalize_nontoken_characters}{}

\item{site_code_col}{}

\item{alt_site_code}{}

\item{sampling_type}{}
}
\description{
read in csv data that meets minimum criteria, built to be as robust
as possible to heterogeneous data formats and contents.
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
