% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_scale_flux_by_area.R
\name{ms_scale_flux_by_area}
\alias{ms_scale_flux_by_area}
\title{Scale stream or precipitation chemical flux by watershed area}
\usage{
ms_scale_flux_by_area(d, site_data)
}
\arguments{
\item{d}{\code{data.frame}. A \code{data.frame} in MacroSheds format (see \href{https://doi.org/10.6084/m9.figshare.c.5621740}{MacroSheds dataset} for definitions),
containing stream or precipitation flux data. A \code{data.frame} in
MacroSheds format is generated by \code{\link[=ms_load_product]{ms_load_product()}}.}

\item{site_data}{\code{data.frame}. Optional; if supplied, should be the MacroSheds
site_data tibble as returned by \code{\link[=ms_download_site_data]{ms_download_site_data()}}. If omitted,
\code{\link[=ms_download_site_data]{ms_download_site_data()}} will be called for you.}
}
\value{
returns a \code{data.frame} in MacroSheds format with stream or precipitation chemical flux in kg/ha/d.
}
\description{
MacroSheds chemical flux is scaled to watershed area by default (kg/ha/d).
This function converts chemical flux to kg/d. If You generated precipitation chemical
flux yourself with \code{\link[=ms_calc_watershed_precip]{ms_calc_watershed_precip()}}, it is in kg/d and this function may be needed.
}
\details{
MacroSheds reports all chemical flux on a per area basis (kg/ha/d) to
facilitate comparison of watersheds of different sizes. If you've unscaled
MacroSheds flux data with \code{\link[=ms_undo_scale_flux_by_area]{ms_undo_scale_flux_by_area()}}, or generated
precipitation flux with \code{\link[=ms_calc_watershed_precip]{ms_calc_watershed_precip()}}, you can use this function
to convert from kg/d to kd/ha/d. To download MacroSheds data, see \code{\link[=ms_download_core_data]{ms_download_core_data()}}.
}
\examples{
### Load some MacroSheds data:
ms_root = 'data/macrosheds'
ms_download_core_data(macrosheds_root = ms_root,
                      domains = 'hbef')
d <- ms_load_product(macrosheds_root = ms_root, 
                     prodname = 'stream_flux_inst_scaled', 
                     domains = 'hbef',
                     filter_vars = 'NO3_N')

d <- ms_undo_scale_flux_by_area(d)
d <- ms_scale_flux_by_area(d)
}
\seealso{
\code{\link[=ms_undo_scale_flux_by_area]{ms_undo_scale_flux_by_area()}}
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
