% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_separate_baseflow.R
\name{ms_separate_baseflow}
\alias{ms_separate_baseflow}
\title{Calculate baseflow}
\usage{
ms_separate_baseflow(discharge, alpha = 0.975, n.reflected = 30)
}
\arguments{
\item{discharge}{\code{data.frame}. A MacroSheds discharge table of one site}

\item{alpha}{Numeric. Default 0.975, alpha value for the Lynne-Hollick baseflow filter. See
details for more information.}

\item{n.reflected}{Numeric. Default 30, see hydrostats::baseflow documentation for more information.}
}
\value{
returns a discharge \code{data.frame} in MacroSheds format with the added column of "baseflow"
}
\description{
Use the Lynne-Hollick baseflow filter, via hydrostats, to perform baseflow separation
}
\details{
This function uses the Lynne-Hollick baseflow filter to separate baseflow
and stormflow. The underlying package hydrostats is used for this calculation. Additional
information on alpha values and n.reflected can be found using ?hydrostats::baseflows
}
\examples{
q = macrosheds::ms_load_product(
    macrosheds_root = 'my/ms/root/',
    prodname = 'discharge',
    domains = 'hbef')
ms_separate_baseflow(q)
}
\author{
Spencer Rhea

Mike Vlah, \email{vlahm13@gmail.com}

Wes Slaughter
}
