% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_data.R
\docType{data}
\name{ms_site_data}
\alias{ms_site_data}
\title{MacroSheds site metadata}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 340 rows and 18 columns.
}
\usage{
ms_site_data
}
\description{
MacroSheds site metadata
}
\author{
Mike Vlah \email{vlahm13@gmail.com}
}
\keyword{data}
