% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_undo_scale_flux_by_area.R
\name{ms_undo_scale_flux_by_area}
\alias{ms_undo_scale_flux_by_area}
\title{Un-scale stream or precipitation chemical flux by watershed area}
\usage{
ms_undo_scale_flux_by_area(d)
}
\arguments{
\item{d}{\code{data.frame}. A \code{data.frame} in MacroSheds format (see details),
containing stream_flux_inst_scaled or precip_flux_inst_scaled data. A \code{data.frame} in
MacroSheds format is generated by \code{\link[=ms_load_product]{ms_load_product()}}.}
}
\value{
returns a \code{data.frame} in MacroSheds format with stream or precipitation chemical flux in kg/d.
}
\description{
MacroSheds chemical flux is scaled to watershed area by default (kg/ha/d).
This function converts chemical flux to kg/d. If You generated precipitation chemical
flux yourself with \code{\link[=ms_calc_watershed_precip]{ms_calc_watershed_precip()}}, it is in kd/d and this function is not needed.
}
\details{
MacroSheds reports all chemical flux on a per area basis (kg/ha/d) to
facilitate comparison of watersheds of different sizes. Sometimes un-scaled
fluxes are of interest. This function will convert the units of a native
MacroSheds chemical flux dataset from kg/ha/d to kg/d. To download
MacroSheds data, see \code{\link[=ms_download_core_data]{ms_download_core_data()}}.

MacroSheds format (only site_code and val columns required as inputs to this function):\tabular{ll}{
   header value \tab column_definition \cr
   date \tab Date in YYYY-mm-dd \cr
   site_code \tab A unique identifier for each MacroSheds site, identical to primary source site code where possible. See \code{\link[=ms_load_sites]{ms_load_sites()}}. \cr
   grab_sample \tab Boolean integer indicating whether the observation was obtained via grab sample or installed sensor. 1 = TRUE (grab sample), 0 = FALSE (installed sensor). \cr
   var \tab Variable code. See \code{\link[=ms_load_variables]{ms_load_variables()}}. \cr
   val \tab Data value. See \code{\link[=ms_load_variables]{ms_load_variables()}} for units. \cr
   ms_status \tab Boolean integer. 0 = clean value. 1 = questionable value. See "Technical Validation" section of \href{https://aslopubs.onlinelibrary.wiley.com/doi/full/10.1002/lol2.10325}{the MacroSheds data paper} for details. \cr
   ms_interp \tab Boolean integer. 0 = measured or imputed by primary source. 1 = interpolated by MacroSheds. See "Temporal Imputation and Aggregation" section of \href{https://aslopubs.onlinelibrary.wiley.com/doi/full/10.1002/lol2.10325}{the MacroSheds data paper} for details. \cr
   val_err \tab The combined standard uncertainty associated with the corresponding data point, if estimable. See "Detection Limits and Propagation of Uncertainty" section of \href{https://aslopubs.onlinelibrary.wiley.com/doi/full/10.1002/lol2.10325}{the MacroSheds data paper} for details. \cr
}
}
\examples{
ms_root <- 'data/macrosheds'
ms_download_core_data(macrosheds_root = ms_root,
                      domains = 'hbef')
d <- ms_load_product(macrosheds_root = ms_root,
                     prodname = 'stream_flux_inst_scaled',
                     domains = 'hbef',
                     filter_vars = 'NO3_N')

d <- ms_undo_scale_flux_by_area(d)
}
\seealso{
\code{\link[=ms_scale_flux_by_area]{ms_scale_flux_by_area()}}
}
\author{
Spencer Rhea

Mike Vlah, \email{vlahm13@gmail.com}

Wes Slaughter
}
