% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_data.R
\docType{data}
\name{ms_vars_ts}
\alias{ms_vars_ts}
\title{MacroSheds variable metadata}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 422 rows and 11 columns.
}
\usage{
ms_vars_ts
}
\description{
MacroSheds variable metadata
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
\keyword{data}
