% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poudre_ws}
\alias{poudre_ws}
\title{Watershed Boundary for the Catch la Poudre river}
\format{
A \code{sf} object containing two columns
\describe{
\item{site_code}{name of site}
\item{geometry}{geospatial information for the watershed}
}
}
\usage{
poudre_ws
}
\description{
Data was retrieved using the nhdplusTools package
}
\references{
Blodgett, D., 2019, nhdplusTools: Tools for Accessing and Working with the NHDPlus, https://code.usgs.gov/water/nhdplusTools
}
\keyword{datasets}
