% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_delineate_watershed.R
\docType{data}
\name{spec_buffer_radius_m}
\alias{spec_buffer_radius_m}
\title{Delineate a watershed from a stream location.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
spec_buffer_radius_m
}
\arguments{
\item{lat}{numeric. Represents the latitude of the pour point in decimal degrees
(negative indicates southern hemisphere).}

\item{long}{numeric. Represents the longitude of the pour point in decimal degrees
(negative indicates west of prime meridian).}

\item{crs}{numeric. Represents the coordinate reference system of your starting
coordinates. EPSG codes work here (e.g. the default is 4326 for WGS 84).}

\item{write_dir}{character string. The directory in which to write output shapefile.}

\item{write_name}{character string. the basename of the shapefile components to
be written. e.g. \code{write_name = 'foo'} would produce foo.shp, foo.shx, foo.prj, foo.dbf.
all the spec_ params}

\item{confirm}{logical. Ignored unless all delineation parameters
(the ones that start with "spec_") are supplied.
If TRUE, you will be asked to visually confirm the delineation before it is
written to \code{write_dir}.}

\item{verbose}{logical. Determines the amount of informative messaging during run.}
}
\value{
A list containing the following components:
\itemize{
\item watershed_area_ha: the area of the delineated watershed in hectares
(meters squared divided by 10,000)
\item buffer_radius_m: the width (meters) around the site location that was used when
requesting a DEM (digital elevation model)
\item snap_distance_m: the search radius (meters) around the pour point that was used
to choose a stream to snap the pour point to.
\item snap_method: either "standard", which snaps the pour point to the cell
within snap_distance_m with highest flow accumulation, or "jenson",
which snaps to the nearest flow line
\item dem_resolution: passed to elevatr::get_elev_raster (z parameter).
depends on supplied machine_status
\item flat_increment: see whitebox::wbt_breach_depressions or
whitebox::wbt_breach_depressions_least_cost
\item breach_method: string. Either 'basic', indicating that
whitebox::wbt_breach_depressions was used, or 'lc', indicating
whitebox::wbt_breach_depressions_least_cost (which less readily
alters the DEM)
\item burn_streams: TRUE or FALSE, indicating whether
whitebox::wbt_burn_streams_at_roads and whitebox::wbt_fill_burn were
used on the DEM
}
}
\description{
From any geographic location, this function (iteratively) attempts to delineate a watershed,
i.e. the land area that contributes overland flow to that location. Delineation
can be guided by user knowledge of local geography, or it can be fully agnostic.
Any specifications not explicitly passed are given reasonable defaults. Candidate
delineations are made available for viewing and the user is asked to pick one,
or try again by interactively providing different specifications. See details.
}
\details{
Output files are unprojected (WGS 84), though processing is done
on projected data. A projection is chosen automatically by \code{\link[=choose_projection]{choose_projection()}},
based on pour point location. Note that this has nothing to do with the crs parameter,
which only allows you to specify the coordinate reference system of your input
coordinates. Also note that for watersheds that span several latitudes or longitudes,
calculated watershed areas might be inaccurate.

For the fully agnostic delineation procedure, here are the steps:
\enumerate{
\item A reasonable projection is chosen via \code{\link[=choose_projection]{choose_projection()}}.
\item A digital elevation model is retrieved via \code{\link[elevatr:get_elev_raster]{elevatr::get_elev_raster()}}.
This initial DEM is 4 km^2 and is centered on the given location.
\item Single-cell pits in the DEM are filled using \code{\link[whitebox:wbt_fill_single_cell_pits]{whitebox::wbt_fill_single_cell_pits()}}.
and larger depressions are breached using \code{\link[whitebox:wbt_breach_depressions]{whitebox::wbt_breach_depressions()}}.
These preprocessing steps ensure that the DEM is traversable by the delineator.
If you'd like to see a "fill depressions" option, please open a
\href{https://github.com/MacroSHEDS/macrosheds/issues}{GitHub issue}.
\item If specified, streamline and roadway files are retrieved via \pkg{osmdata},
and streamlines are \emph{burned} into the DEM at stream-road intersections.
In most cases, this allows delineation to proceed through bridges and culverts
where it would otherwise stop.
\item Flow direction and accumulation are computed via \code{\link[whitebox:wbt_d8_pointer]{whitebox::wbt_d8_pointer()}}
and \code{\link[whitebox:wbt_d8_flow_accumulation]{whitebox::wbt_d8_flow_accumulation()}}.
\item The specified site location is \emph{snapped} to a nearby stream line in
multiple ways, in the hopes that at least one snap will land on the
intended stream. This matters when e.g. a tributary is being delineated
near its confluence with a mainstem. Snapping methods include
\code{\link[whitebox:wbt_snap_pour_points]{whitebox::wbt_snap_pour_points()}} and \code{\link[whitebox:wbt_jenson_snap_pour_points]{whitebox::wbt_jenson_snap_pour_points()}}.
\item All unique snap locations are delineated via \code{\link[whitebox:wbt_watershed]{whitebox::wbt_watershed()}}. If
at any point the computed watershed meets the edge of the DEM, a larger DEM
is retrieved, and the delineator tries again.
\item All candidate watersheds are presented for inspection. Either one is selected,
or the procedure starts over with a new set of user-given and/or default
specifications.
\item The delineated watershed is saved as a shapefile. Watershed area
and the successful set of specifications are returned.
}
}
\examples{
area_and_specs <- ms_delineate_watershed(
    lat = 44.21013,
    long = -122.2571,
    crs = 4326,
    write_dir = '/some/path',
    write_name = 'example_site'
)
}
\seealso{
\code{\link[=ms_scale_flux_by_area]{ms_scale_flux_by_area()}}, \code{\link[=ms_undo_scale_flux_by_area]{ms_undo_scale_flux_by_area()}}
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}

Spencer Rhea

Wes Slaughter
}
\keyword{datasets}
