% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_data.R
\docType{data}
\name{ts_bib}
\alias{ts_bib}
\title{MacroSheds references as BibTeX strings}
\format{
An object of class \code{character} of length 1.
}
\usage{
ts_bib
}
\description{
MacroSheds references as BibTeX strings
}
\author{
Mike Vlah \email{vlahm13@gmail.com}
}
\keyword{data}
