% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_calc_flux.R
\name{ms_calc_flux}
\alias{ms_calc_flux}
\title{Calculate daily solute fluxes}
\usage{
ms_calc_flux(chemistry, q, verbose = TRUE, ws_area_scaled = TRUE)
}
\arguments{
\item{chemistry}{\code{data.frame}. A \code{data.frame} of precipitation or
stream chemistry data in MacroSheds format and in units of mg/L.}

\item{q}{\code{data.frame}. A \code{data.frame} of precipitation or stream
discharge in MacroSheds format and in units of mm or L/s, respectively.}

\item{verbose}{logical. Default TRUE; prints more information to console.}

\item{ws_area_scaled}{logical. Default TRUE; flux values are divided by watershed area.}
}
\value{
returns a \code{tibble} of stream or precipitation chemical flux for every timestep
where discharge/precipitation and chemistry are reported. Output units are kg/ha/timestep.
}
\description{
Determines solute fluxes from interpolated daily Q (discharge
or precipitation) and corresponding chemistry data.
}
\details{
For monthly and annual fluxes computed by various methods, see \code{\link[=ms_calc_flux_rsfme]{ms_calc_flux_rsfme()}}.

Here, solute flux is calculated by multiplying solute concentration by flow
of water (flux = concentration * flow). The output units depend on the time
interval at which input data are collected. The resulting flux units will always be
kg/ha/T, where T is the time interval of the input.
Stream and precipitation flux are calculated differently. For stream flux (stream chemistry * discharge),
flux is calculated as: kg/ha/T = mg/L * L/s * T / 1e6 / ws_area.
For precipitation flux (precip chemistry * precip depth over the watershed) is calculated as: kg/ha/T = mg/L * mm/T / 100.
You can convert between kg/ha/T and kg/T using \code{\link[=ms_scale_flux_by_area]{ms_scale_flux_by_area()}} and
\code{\link[=ms_undo_scale_flux_by_area]{ms_undo_scale_flux_by_area()}}, which can also be applied using the ws_area_scaled parameter to
this function.

Before running \code{\link[=ms_calc_flux]{ms_calc_flux()}}, ensure both \code{q} and
\code{chemistry} have the same time interval. See \code{\link[=ms_synchronize_timestep]{ms_synchronize_timestep()}}.
Also ensure chemistry units are mg/L. See \code{\link[=ms_conversions]{ms_conversions()}}.
}
\examples{
#' ### Load some MacroSheds data:
ms_root = 'data/macrosheds'
ms_download_core_data(macrosheds_root = ms_root,
                      domains = 'hbef')
chemistry <- ms_load_product(macrosheds_root = ms_root,
                             prodname = 'stream_chemistry',
                             site_codes = c('w1', 'w3', 'w6'),
                             filter_vars = c('NO3_N', 'Cl', 'Na'))

q <- ms_load_product(macrosheds_root = ms_root,
                     prodname = 'discharge',
                     site_codes = c('w1', 'w3', 'w6'))

flux <- ms_calc_flux(chemistry = chemistry,
                     q = q)
}
\seealso{
\code{\link[=ms_calc_flux_rsfme]{ms_calc_flux_rsfme()}}, \code{\link[=ms_synchronize_timestep]{ms_synchronize_timestep()}}, \code{\link[=ms_conversions]{ms_conversions()}}, \code{\link[=ms_scale_flux_by_area]{ms_scale_flux_by_area()}}, \code{\link[=ms_undo_scale_flux_by_area]{ms_undo_scale_flux_by_area()}}
}
\author{
Spencer Rhea

Mike Vlah, \email{vlahm13@gmail.com}

Wes Slaughter
}
