% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_catalog.R
\name{ms_catalog}
\alias{ms_catalog}
\title{Download macrosheds variable catalog (deprecated)}
\usage{
ms_catalog()
}
\description{
this function has been retired, and replaced with \code{ms_load_variables}, please use this instead
}
\seealso{
\code{\link[=ms_load_variables]{ms_load_variables()}}
}
\author{
Wes Slaughter, \email{wslaughter@berkeley.edu}

Mike Vlah

Spencer Rhea
}
