% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_catalog.R
\name{ms_catalog}
\alias{ms_catalog}
\title{Query MacroSheds variable catalog}
\usage{
ms_catalog(fp = NULL)
}
\value{
returns a \code{data.frame} of the variables in
the MacroSheds dataset and metadata about them.
}
\description{
downloads the MacroSheds variable catalog, which documents
important metadata for all variables in MacroSheds -- such
as sites, number of observations, first and last record of
at each site, and more.
}
\examples{
macrosheds_var_catalog <- ms_download_var_catalog()
}
\seealso{
\code{\link[=ms_download_variables]{ms_download_variables()}}, \code{\link[=ms_download_site_data]{ms_download_site_data()}}
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
