% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_delineate_watershed.R
\name{ms_delineate_watershed}
\alias{ms_delineate_watershed}
\title{Delineate a watershed from a stream location.}
\usage{
ms_delineate_watershed(lat, long, crs, write_dir, write_name, verbose = TRUE)
}
\arguments{
\item{d}{\code{data.frame}. A macrosheds \code{dataframe} downloaded using
\code{ms_downloadcore_data()}}

\item{desired_interval}{character. Input that can be parsed by the "by"
parameter to base::seq.POSIXt, e.g. "5 mins", "1 day", "1 month"}

\item{impute_limit}{numeric. The maximum number of consecutive points to
inter/extrapolate. it's passed to imputeTS::na_interpolate.}

\item{summary_fun}{One of "mean" or "sum". The function that will be used to
summarize data if desired_interval is larger than the input data in d}

\item{quiet}{logical. Should warnings be printed to console}
}
\value{

}
\description{
From any geographic location, this function (iteratively) attempts to delineate a watershed,
i.e. the land area that contributes overland flow to that location. Delineation
can be guided by user knowledge of local geography, or it can be fully agnostic.
Any specifications not explicitly passed are given reasonable defaults. Unless
\code{confirm} is set to \code{FALSE}, candidate delineations are made available
for viewing and the user is asked to pick one or try again with different
specifications.
}
\details{
For the fully agnostic procedure, here are the steps:
\enumerate{
\item A reasonable projection is chosen via \code{\link[=choose_projection]{choose_projection()}}.
\item A digital elevation model is retrieved via \code{\link[elevatr:get_elev_raster]{elevatr::get_elev_raster()}}.
This initial DEM is 4 km^2 and is centered on the given location.
\item Single-cell pits in the DEM are filled using \code{\link[whitebox:wbt_fill_single_cell_pits]{whitebox::wbt_fill_single_cell_pits()}}.
and larger depressions are breached using \code{\link[whitebox:wbt_breach_depressions]{whitebox::wbt_breach_depressions()}}.
These preprocessing steps ensure that the DEM is traversible by the delineator.
If you'd like to see a "fill depressions" option, please open a
\href{https://github.com/MacroSHEDS/macrosheds/issues}{GitHub issue}.
\item If specified, streamline and roadway files are retrieved via \pkg{osmdata},
and streamlines are \emph{burned} into the DEM at stream-road intersections.
In most cases, this allows delineation to proceed through bridges and culverts.
\item Flow direction and accumulation are computed via \code{\link[whitebox:wbt_d8_pointer]{whitebox::wbt_d8_pointer()}}
and \code{\link[whitebox:wbt_d8_flow_accumulation]{whitebox::wbt_d8_flow_accumulation()}}.
\item The specified site location is \emph{snapped} to a nearby stream line in
multiple ways, in the hopes that at least one snap will land on the correct
intended stream. This matters when e.g. a tributary is being delineated
near its confluence with a mainstem. Snapping methods include
\code{\link[whitebox:wbt_snap_pour_points]{whitebox::wbt_snap_pour_points()}} and \code{\link[whitebox:wbt_jenson_snap_pour_points]{whitebox::wbt_jenson_snap_pour_points()}}.
\item All unique snap locations are delineated via \code{\link[whitebox:wbt_watershed]{whitebox::wbt_watershed()}}. If
at any point the computed watershed meets the edge of the DEM, a larger DEM
is retrieved, and the delineator tries again.
\item All candidate watersheds are presented for inspection. Either one is selected,
or the procedure starts over with a new set of user-given and/or default
specifications.
\item The delineated watershed is saved as a shapefile. Watershed area
and the successful set of specifications are returned.
}

ms_synchronize_timestep is used to convert data into a finer
temporal resolution, such as taking a weekly grab sample and interpolating
the values to have daily data, or to aggregate data into large intervals
such as monthly and annual sums or means. Macrosheds data undergoes temporal
synchronization in our data processing pipeline. Our data is aggregated/imputed
to a daily time step for all variables and the impute limit for discharge and
precipitation is 3 days and for chemistry variables 15 days.
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}

Spencer Rhea

Wes Slaughter
}
