% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_download_core_data.R
\name{ms_download_core_data}
\alias{ms_download_core_data}
\title{Download macrosheds core datasets}
\usage{
ms_download_core_data(
  macrosheds_root,
  networks,
  domains,
  version = "1.0",
  quiet = FALSE
)
}
\arguments{
\item{macrosheds_root}{character. Directory where macrosheds data files will be downloaded.
If this directory does not exist, it will be created.}

\item{networks}{character vector. macrosheds networks that will be downloaded.
Either a single network, vector of networks, or 'all'. See \code{ms_load_sites()}
for networks available for download.}

\item{domains}{character vector. macrosheds domains that will be downloaded.
Either a single domain, vector of domains, or 'all'. See \code{ms_load_sites()}
for domains available for download.}

\item{version}{character. The MacroSheds dataset version to download, e.g. "1.0". Defaults to
most recent. As of 2023-03-17, only version 1.0 is available, so this parameter is a stub.}

\item{quiet}{logical. If TRUE, some messages will be suppressed.}
}
\value{
Downloads all core data for selected domains to the
directory specified by \code{macrosheds_root}. Site datasets are arranged according to the following
structure: domain/prodname/site_code.feather. For definitions of these terms as used by
MacroSheds, see \href{https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=1262}{MacroSheds EDI}.
}
\description{
Download the macrosheds core datasets of stream chemistry, stream
discharge, stream chemical flux, precipitation, precipitation chemistry,
precipitation chemical flux, ws_boundary, stream_gauge_locations, and
precip_gauge_locations. Not all products are available at all sites, but
all products available for the selected domains will be downloaded.
}
\details{
Either \code{networks} or \code{domains} must be supplied. If 'all' is
supplied to either argument, all domains will be downloaded regardless of
what is supplied to the other argument. The full (core) dataset is approximately 500 MiB when
compressed and may take minutes to hours to download, depending on connection speed. When uncompressed,
the dataset is approximately 4 GiB, so be mindful of disk space.
}
\examples{
dir.create('data/macrosheds', recursive = TRUE)
ms_download_core_data(macrosheds_root = 'data/macrosheds',
                      domains = c('niwot', 'hjandrews'))
}
\seealso{
\code{\link[=ms_load_sites]{ms_load_sites()}}, \code{\link[=ms_load_variables]{ms_load_variables()}}, \code{\link[=ms_load_product]{ms_load_product()}}, \code{\link[=ms_load_spatial_product]{ms_load_spatial_product()}}
}
\author{
Spencer Rhea

Mike Vlah, \email{vlahm13@gmail.com}

Wes Slaughter
}
