% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_download_core_data.R
\name{ms_download_core_data}
\alias{ms_download_core_data}
\title{Download macrosheds core datasets}
\usage{
ms_download_core_data(
  macrosheds_root,
  networks,
  domains,
  version = "latest",
  skip_existing = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{macrosheds_root}{character. Directory where macrosheds data files will be downloaded.
If this directory does not exist, it will be created. The \code{macrosheds_root} directory should
only be used for MacroSheds data files.}

\item{networks}{character vector. macrosheds networks that will be downloaded.
Either a single network, vector of networks, or 'all'. Either \code{networks} or \code{domains} must be supplied.
See \code{\link[=ms_load_sites]{ms_load_sites()}} for networks available for download.}

\item{domains}{character vector. macrosheds domains that will be downloaded.
Either a single domain, vector of domains, or 'all'. Either \code{domains} or \code{networks} must be supplied.
See \code{\link[=ms_load_sites]{ms_load_sites()}} for domains available for download.}

\item{version}{numeric or "latest". The MacroSheds dataset version to download.}

\item{skip_existing}{logical. If FALSE, dataset components already downloaded to macrosheds_root will
be overwritten. If TRUE, these will be skipped.}

\item{quiet}{logical. If TRUE, some messages will be suppressed.}
}
\value{
Downloads all core data for selected domains to the
directory specified by \code{macrosheds_root}. Site datasets are arranged according to the following
structure: domain/prodname/site_code.feather. For definitions of these terms as used by
MacroSheds, see glossary.txt on \href{https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=1262}{MacroSheds EDI}.
}
\description{
Download the macrosheds core datasets of stream chemistry, stream
discharge, stream solute flux, precipitation, precipitation chemistry,
precipitation solute flux, annual solute loads, ws_boundary, stream_gauge_locations, and
precip_gauge_locations. Not all products are available at all sites, but
all products available for the selected domains will be downloaded.
}
\details{
The core time-series dataset is approximately 6 GiB uncompressed.
}
\examples{
ms_download_core_data(macrosheds_root = 'data/macrosheds',
                      domains = c('niwot', 'hjandrews'))
}
\seealso{
\code{\link[=ms_load_sites]{ms_load_sites()}}, \code{\link[=ms_load_variables]{ms_load_variables()}}, \code{\link[=ms_load_product]{ms_load_product()}}, \code{\link[=ms_load_spatial_product]{ms_load_spatial_product()}}
}
\author{
Spencer Rhea

Mike Vlah, \email{vlahm13@gmail.com}

Wes Slaughter
}
