% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_download_site_data.R
\name{ms_download_site_data}
\alias{ms_download_site_data}
\title{Download macrosheds site dataset (deprecated)}
\usage{
ms_download_site_data()
}
\description{
this function has been retired, and replaced with \code{ms_load_sites}, please use this insetad
}
\seealso{
\code{\link[=ms_load_sites]{ms_load_sites()}}
}
\author{
Wes Slaughter, \email{wslaughter@berkeley.edu}

Mike Vlah

Spencer Rhea
}
