% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_download_site_data.R
\name{ms_download_site_data}
\alias{ms_download_site_data}
\title{Download MacroSheds site metadata table}
\usage{
ms_download_site_data()
}
\value{
returns a \code{tibble} of metadata for all MacroSheds sites,
including network, domain, location, area, and additional information.
For definitions, see PLACEHOLDER0.
}
\description{
Download the MacroSheds site data table, including information on networks, domains,
stream gauge locations, and more.
}
\examples{
macrosheds_sites <- ms_downloadsite_data()
}
\seealso{
\code{\link[=ms_download_variables]{ms_download_variables()}}, \code{\link[=ms_download_core_data]{ms_download_core_data()}}
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
