% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_download_variables.R
\name{ms_download_variables}
\alias{ms_download_variables}
\title{Download macrosheds variables data table}
\usage{
ms_download_variables()
}
\value{
returns a \code{data.frame} of variables, codes,
and units in MacroSheds format
}
\description{
Download a table of MacroSheds variables with information on variable codes,
units, and more.
}
\examples{
ms_vars <- ms_download_variables()
}
\seealso{
\code{\link[=ms_download_site_data]{ms_download_site_data()}}, \code{\link[=ms_download_core_data]{ms_download_core_data()}}
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
