% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_download_variables.R
\name{ms_download_variables}
\alias{ms_download_variables}
\title{Download macrosheds varibles data table}
\usage{
ms_download_variables()
}
\value{
returns a \code{data.frame} of variables, codes,
    and units in macrosheds
}
\description{
Download the macrosheds variables table with information on variables codes,
units, and more
}
\examples{
macrosheds_sites <- download_ms_variables()
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
