% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_download_ws_attr.R
\name{ms_download_ws_attr}
\alias{ms_download_ws_attr}
\title{Download MacroSheds watershed attribute summary data tables}
\usage{
ms_download_ws_attr(ws_attr_dir, dataset = "summaries", quiet = FALSE)
}
\arguments{
\item{ws_attr_dir}{character. Directory where macrosheds watershed attribute data files will be downloaded.
If this directory does not exist, it will be created.}

\item{dataset}{character. This function can download the two types of watershed attribute dataset
provided by macrosheds. "summaries" will download a feather file containing watershed attributes
summarized for each macrosheds site over all available data. "time series" will download 6 feather
files containing time series data of watershed attributes, these files are each prefixed "spatial_timeseries":
"vegetation", "terrain", "parentmaterial", "landcover", "hydrology", "climate"}

\item{quiet}{logical. If TRUE, some messages will be suppressed.}
}
\value{
returns nothing.
}
\description{
Download MacroSheds watershed attribute data for each site, both summary and time series datasets
are available
}
\examples{
ms_download_ws_attr_summary(ws_attr_dir = 'ms/spatial', dataset = 'time series', quiet = FALSE)
}
\seealso{
\code{\link[=ms_download_site_data]{ms_download_site_data()}}, \code{\link[=ms_download_core_data]{ms_download_core_data()}}
}
\author{
Wes Slaughter, \email{wslaughter@berkeley.edu}

Mike Vlah

Spencer Rhea, \email{spencerrhea41@gmail.com}
}
