% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_generate_attribution.R
\name{ms_generate_attribution}
\alias{ms_generate_attribution}
\title{Easily cite and/or acknowledge primary sources of the MacroSheds dataset}
\usage{
ms_generate_attribution(d, site_data = NULL, attrib_data___________ = NULL)
}
\arguments{
\item{chemistry}{\code{data.frame}. A \code{data.frame} of precipitation or}
}
\value{
returns a
where
}
\description{
Also returns contact information, URLs, DOIs, and intellectual rights
details, all based on an input data.frame in MacroSheds format.
}
\details{
Chemical
\code{\link[=ms_undo_scale_flux_by_area]{ms_undo_scale_flux_by_area()}}.

Before running \code{\link[=ms_calc_flux]{ms_calc_flux()}}, ensure both \code{q} and
}
\examples{
#' ### Load some MacroSheds data:
ms_root = 'data/macrosheds'
}
\seealso{
\code{\link[=ms_synchronize_timestep]{ms_synchronize_timestep()}},
}
\author{
Mike Vlah, vlahm13@gmail.com

Wes Slaughter
}
