% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_identify_usgs_gauges.R
\name{ms_identify_usgs_gauges}
\alias{ms_identify_usgs_gauges}
\title{Identify USGS gauges near a point}
\usage{
ms_identify_usgs_gauges(sites, lat = NULL, long = NULL, radius = 500)
}
\arguments{
\item{sites}{Either an \code{sf} object with locations of interest or a \code{data.frame}
containing latitude and longitude information.}

\item{lat}{Character. If sites is not a \code{sf} object, the name of the column
containing latitudes in WGS84. Default NULL.}

\item{long}{Character. If sites is not a \code{sf} object, the name of the column
containing longitudes in WGS84. Default NULL.}

\item{radius}{Numeric. The radius in meters within which to search for USGS gauges; Default 500.}
}
\value{
returns a \code{data.frame} with the column \code{usgs_site}.
}
\description{
Search for USGS gauges within a radius of a location specified by latitude and longitude
}
\details{
This function will search for a USGS gauge within the radius supplied
for every site in the sites file. The \code{dataRetrieval} package is used to locate USGS gauges.
If no USGS gauge is found, then the \code{usgs_site} column will be NA for that site.
}
\examples{
# With lat long 
sites <- tibble(site_code = c('MA_AE03', 'MI_KR01'),
                Latitude = c(42.04856, 42.28575),
                Longitude = c(-72.45403, -85.51467))
ms_identify_usgs_gauges(sites = sites, lat = 'Latitude', long = 'Longitude')

# With sf
site_sf <- sites \%>\%
    sf::st_as_sf(coords = c('long', 'lat'), crs = 4326) \%>\%
    sf::st_buffer(radius)
}
\author{
Spencer Rhea, \email{spencerrhea41@@gmail.com}

Mike Vlah

Wes Slaughter
}
