% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_load_product.R
\name{ms_load_product}
\alias{ms_load_product}
\title{Load in macrosheds data products}
\usage{
ms_load_product(
  macrosheds_root,
  prodname,
  filter_vars,
  networks,
  domains,
  site_codes,
  sort_result = FALSE,
  warn = TRUE
)
}
\arguments{
\item{macrosheds_root}{character. The path to the macrosheds dataset's parent
directory, e.g. '~/stuff/macrosheds_dataset_v0.3'.}

\item{prodname}{character. read and combine files associated with this prodname. 
Available prodnames are:
discharge, stream_chemistry, stream_flux_inst, precipitation,
precip_chemistry, precip_flux_inst.}

\item{filter_vars}{character vector. for products like stream_chemistry that include
multiple variables, this filters to just the ones specified (ignores
variable prefixes). To see a catalog of variables, visit macrosheds.org.}

\item{networks}{character vector. Networks to load, optional.}

\item{domains}{character vector. Domains to load, optional.}

\item{site_codes}{character vector. Sites to load, optional.}

\item{sort_result}{logical. If TRUE, output will be sorted by site_code, var,
datetime. this may take a few additional minutes for some products in
the entire dataset.}

\item{warn}{logical. If TRUE, function will give a prompt with the estimated 
file size before loading in data.}
}
\value{
returns a \code{tibble} containing all the all data belonging to the 
   selected product and variables in the \code{macrosheds_root} directory
}
\description{
Loads and filters macrosheds products (e.g stream_chemistry, discharge, etc.) from a 
downloaded macrosheds dataset see \code{download_ms_code_data()} for data download.
}
\examples{
macrosheds_data <- load_product(macrosheds_root = 'data/macrosheds_v1', 
                                prodname = 'stream_chemistry', 
                                sort_result = FALSE, 
                                filter_vars = 'NO3_N')
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
