% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_load_product.R
\name{ms_load_product}
\alias{ms_load_product}
\title{Read MacroSheds data products from disk into R}
\usage{
ms_load_product(
  macrosheds_root,
  prodname,
  filter_vars,
  networks,
  domains,
  site_codes,
  sort_result = FALSE,
  warn = TRUE
)
}
\arguments{
\item{macrosheds_root}{character. The path to the MacroSheds dataset's parent
directory, established with \code{\link[=ms_download_core_data]{ms_download_core_data()}}.}

\item{prodname}{character. A MacroSheds product name. Files associated with this
product name will be read and combined. Available prodnames are:
discharge, stream_chemistry, stream_flux_inst, precipitation,
precip_chemistry, precip_flux_inst.}

\item{filter_vars}{character vector. for products like stream_chemistry that include
multiple variables, this filters to just the ones specified (ignores
variable prefixes). To see a catalog of variables, visit macrosheds.org or see \href{https://doi.org/10.6084/m9.figshare.c.5621740}{MacroSheds documentation}.}

\item{networks}{character vector. MacroSheds networks to load; optional. To see a catalog of
networks, visit macrosheds.org or see \href{https://doi.org/10.6084/m9.figshare.c.5621740}{MacroSheds documentation}.}

\item{domains}{character vector. MacroSheds domains to load; optional. To see a catalog of
domains, visit macrosheds.org or see \href{https://doi.org/10.6084/m9.figshare.c.5621740}{MacroSheds documentation}.}

\item{site_codes}{character vector. MacroSheds sites to load, optional. To see a catalog of
site_codes, visit macrosheds.org or see \href{https://doi.org/10.6084/m9.figshare.c.5621740}{MacroSheds documentation}.}

\item{sort_result}{logical. If TRUE, output will be sorted by site_code, var,
datetime. this may add considerable loading time for large datasets.}

\item{warn}{logical. If TRUE, function will not load more than 100MB without permission.}
}
\value{
Returns a \code{tibble} in MacroSheds format. See \href{https://doi.org/10.6084/m9.figshare.c.5621740}{MacroSheds documentation} for definitions.
}
\description{
Loads and optionally filters MacroSheds time-series products (e.g stream_chemistry, discharge, etc.) from a
downloaded MacroSheds dataset. For watershed boundaries and gauge locations, see \code{\link[=ms_load_spatial_product]{ms_load_spatial_product()}}.
}
\examples{
ms_root = 'data/macrosheds'
dir.create(ms_root, recursive = TRUE)
ms_download_core_data(macrosheds_root = ms_root,
                      domains = c('niwot', 'hjandrews'))
macrosheds_data <- ms_load_product(macrosheds_root = ms_root, 
                                   prodname = 'stream_chemistry', 
                                   filter_vars = 'NO3_N')
}
\seealso{
\code{\link[=ms_download_core_data]{ms_download_core_data()}}, \code{\link[=ms_load_spatial_product]{ms_load_spatial_product()}}, \code{\link[=ms_load_variables]{ms_load_variables()}}, \code{\link[=ms_load_sites]{ms_load_sites()}}
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
