% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_load_product.R
\name{ms_load_product}
\alias{ms_load_product}
\title{Read MacroSheds data products from disk into R}
\usage{
ms_load_product(
  macrosheds_root,
  prodname,
  filter_vars,
  networks,
  domains,
  site_codes,
  sort_result = FALSE,
  warn = TRUE
)
}
\arguments{
\item{macrosheds_root}{character. The path to the MacroSheds dataset's parent
directory, established with \code{\link[=ms_download_core_data]{ms_download_core_data()}} and/or \code{\link[=ms_download_ws_attr]{ms_download_ws_attr()}}.
If you specified different locations with each of these functions, you'll need to
refer to them separately when loading time-series data vs. watershed attribute data.}

\item{prodname}{character. A MacroSheds product name. Files associated with this
product name will be read and combined (see * caveat). Available prodnames are (for core time-series products):
\itemize{
\item discharge
\item stream_chemistry
\item stream_flux*
\item precipitation,
\item precip_chemistry
\item precip_flux*
}

(and for watershed attribute products):
\itemize{
\item ws_attr_summaries
\item ws_attr_timeseries:all
\item ws_attr_timeseries:climate
\item ws_attr_timeseries:hydrology
\item ws_attr_timeseries:landcover
\item ws_attr_timeseries:parentmaterial
\item ws_attr_timeseries:terrain
\item ws_attr_timeseries:vegetation
\item ws_attr_CAMELS_summaries
\item ws_attr_CAMELS_Daymet_forcings

*note that flux products cannot be loaded using
this function, but can be calculated from component products using \code{\link[=ms_calc_flux]{ms_calc_flux()}}
and \code{\link[=ms_calc_flux_rsfme]{ms_calc_flux_rsfme()}}.
}}

\item{filter_vars}{character vector. for products like stream_chemistry that include
multiple variables, this filters to just the ones specified (ignores
variable prefixes). Ignored if requesting discharge, precipitation, or watershed attributes.
To see a catalog of variables, visit macrosheds.org or see \href{https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=1262}{MacroSheds EDI}.}

\item{networks, domains, site_codes}{character vectors of MacroSheds networks/domains/sites to load. Omit to load all. See \code{\link[=ms_load_sites]{ms_load_sites()}},
or visit \link{macrosheds.org} or \href{https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=1262}{EDI}.}

\item{sort_result}{logical. Ignored if requesting watershed attributes.
If TRUE, and requesting core time-series data, output will be sorted by site_code, var,
datetime. this may add considerable loading time for large datasets.}

\item{warn}{logical. If TRUE, function will not load more than 100 MB without permission.}
}
\value{
Returns a \code{tibble} in MacroSheds format. See \href{https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=1262}{MacroSheds EDI} for definitions.
}
\description{
Loads and optionally filters MacroSheds time-series products (e.g stream_chemistry, discharge, etc.) from a
downloaded MacroSheds dataset. For watershed boundaries and gauge locations, see \code{\link[=ms_load_spatial_product]{ms_load_spatial_product()}}.
}
\examples{
ms_root = 'data/macrosheds'
ms_download_core_data(macrosheds_root = ms_root, domains = c('niwot', 'hjandrews'))
ms_download_ws_attr(macrosheds_root = ms_root, dataset = 'time series')
macrosheds_data <- ms_load_product(macrosheds_root = ms_root, 
                                   prodname = 'stream_chemistry', 
                                   filter_vars = 'NO3_N')
macrosheds_data <- ms_load_product(macrosheds_root = ms_root, 
                                   prodname = 'ws_attr_timeseries:all')
}
\seealso{
\code{\link[=ms_download_core_data]{ms_download_core_data()}}, \code{\link[=ms_download_ws_attr]{ms_download_ws_attr()}}, \code{\link[=ms_load_spatial_product]{ms_load_spatial_product()}}, \code{\link[=ms_load_variables]{ms_load_variables()}}, \code{\link[=ms_load_sites]{ms_load_sites()}}
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
