% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_load_variables.R
\name{ms_load_variables}
\alias{ms_load_variables}
\title{load MacroSheds variables data table}
\usage{
ms_load_variables(var_set = "timeseries")
}
\arguments{
\item{var_set}{character. One of "timeseries", "timeseries_by_site", or "ws_attr".
If "timeseries", basic metadata for core MacroSheds time series variables will be returned. If "timeseries_by_site", a
more complete catalog of variable availability per site will be returned. If "ws_attr", metadata for MacroSheds watershed
attributes will be returned.}
}
\value{
returns a \code{data.frame} of variables, codes, units, and other metadata.
}
\description{
Load a data.frame of MacroSheds variables with information on variable codes,
units, and more. Visit \href{https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=1262}{EDI} for complete metadata.
}
\details{
Variable metadata are included as an .RData file with the macrosheds package, and
are loaded into the macrosheds namespace on library(macrosheds). This
function simply retrieves a data.frame from the macrosheds namespace and makes it available
in the global environment.
}
\examples{
ws_attr_vars <- ms_load_variables(var_set = 'ws_attr')
}
\seealso{
\code{\link[=ms_load_sites]{ms_load_sites()}}, \code{\link[=ms_load_spatial_product]{ms_load_spatial_product()}}, \code{\link[=ms_load_product]{ms_load_product()}}
}
\author{
Wes Slaughter, \email{weston.slaughter@duke.edu}

Mike Vlah, \email{vlahm13@gmail.com}

Spencer Rhea
}
