% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_run_egret.R
\name{ms_run_egret}
\alias{ms_run_egret}
\title{Run the WRTDS model through EGRET for MacroSheds data}
\usage{
ms_run_egret(
  stream_chemistry,
  discharge,
  prep_data = TRUE,
  run_egret = TRUE,
  kalman = FALSE,
  quiet = FALSE,
  site_data = NULL
)
}
\arguments{
\item{stream_chemistry}{\code{data.frame} in MacroSheds format (see \href{https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=1262}{MacroSheds EDI}
Must contain only one stream chemistry variable and one site_code to comply with WRTDS.
A \code{tibble} of stream chemistry, in
MacroSheds format, can be generated by \code{\link[=ms_load_product]{ms_load_product()}}.}

\item{discharge}{\code{data.frame} in MacroSheds format (see \href{https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=1262}{MacroSheds EDI}
Must contain discharge for only one site_code to comply with WRTDS.
A \code{tibble} of discharge, in
MacroSheds format, can be generated by \code{\link[=ms_load_product]{ms_load_product()}}.}

\item{prep_data}{logical. Should data be thinned/altered to follow EGRET recommendations?
See details for more information.}

\item{run_egret}{logical. If FALSE, the EGRET model will not be run and an EGRET
eList will be returned. This argument is intended for those who want to change
default parameters in the EGRET model but need data in the eList format.}

\item{quiet}{logical. Should warnings be printed to console?}

\item{site_data}{tibble. Default NULL, if you are using a non-MacroSheds site you must
supply a tibble with the columns: site_code, ws_area_ha, latitude, longitude. If
using a MacroSheds site you can leave this argument NULL and the MacroSheds site_data
table will be retrieved.}

\item{Kalman}{logical. Should EGRET run the Kalman filter on WRTDS results? See
\code{\link[EGRET:wrtdsK]{EGRET::WRTDSKalman()}} for more information.}
}
\value{
returns a \code{list} in the EGRET format with model results.
}
\description{
ms_run_egret prepares data for the EGRET package and runs the Weighted Regression
on Time, Discharge, and Season (WRTDS) model.
}
\details{
WRTDS is a model to predict stream chemistry at a daily time step
based on the season, discharge, and time since last grab sample. This model is
useful for calculating flux when sampling is sparse (greater than a weekly
sampling frequency). For more information on the WRTDS and EGRET package
see \url{https://github.com/USGS-R/EGRET}.

\code{prep_data} will remove NAs from the dataset, set all 0 values in stream
chemistry dataset to the minimum value in the dataset (this is required because EGRET will
fail if there are 0 values in a dataset), remove years with fewer than 6 chemistry samples,
and remove datetimes when there is chemistry but no discharge data reported.
}
\seealso{
\code{\link[=ms_load_product]{ms_load_product()}}, \code{\link[=ms_conversions]{ms_conversions()}}
}
\author{
Spencer Rhea

Mike Vlah, \email{vlahm13@gmail.com}

Wes Slaughter
}
\keyword{internal}
