% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_run_egret.R
\name{ms_run_egret}
\alias{ms_run_egret}
\title{Run the WRTDS model through EGRET for macrosheds data}
\usage{
ms_run_egret(
  stream_chemistry,
  discharge,
  prep_data,
  Kalman = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{stream_chemistry}{\code{date.frame}. A macrosheds \code{data.frame} 
containg only one stream chemistry variable and one site_code; WRTDS can 
only be run for one chemistry vatiable and one site at a time. See 
\code{download_ms_core_data()} to download data}

\item{discharge}{\code{date.frame}. A macrosheds \code{data.frame} containing 
discharge data for a given site. See  \code{download_ms_core_data()} to 
download data}

\item{prep_data}{logical. Should data be thined/altered to folow EGRET reomendations. 
See details for more information}

\item{Kalman}{logical. Should EGRET run the Kalman filter on WRTDS results. See
?EGRET::WRTDSKalman() for more information}

\item{quiet}{logical. Should warnings be printed to console}
}
\value{
returns a \code{list} in the EGRET format with modle resutls
}
\description{
ms_run_egret prepairs data for the egret package and runs the Weighted Regretion
on Time, Discharge, and Season (WRTDS) model
}
\details{
WRTDS is a model to predict stream chemistry at a daily time step 
    based on the season, discharge, and time since last grab sample. This model is 
    usfule for calculating flux when sampling is sparce (greater than a weekly 
    sampling frequency). For more informaiton on the WRTDS and EGRET package 
    see https://github.com/USGS-R/EGRET
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
