% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_synchronize_timestep.R
\name{ms_synchronize_timestep}
\alias{ms_synchronize_timestep}
\title{summarize macrosheds data to specified time intervals}
\usage{
ms_synchronize_timestep(
  d,
  desired_interval,
  impute_limit,
  summary_fun,
  quiet = FALSE
)
}
\arguments{
\item{d}{\code{date.frame}. A macrosheds \code{dataframe} downloaded using 
\code{download_ms_core_data()}}

\item{desired_interval}{character. Input that can be parsed by the "by"
parameter to base::seq.POSIXt, e.g. "5 mins", "1 day", "1 month"}

\item{impute_limit}{numeric. The maximum number of consecutive points to
inter/extrapolate. it's passed to imputeTS::na_interpolate.}

\item{summary_fun}{One of "mean" or "sum". The function that will be used to 
summarize data if desired_interval is larger than the input data in d}

\item{quiet}{logical. Should warnings be printed to console}
}
\value{
returns a \code{tibble} with data in the new desired_interval with a 
    ms_interp column indicating if a value was interpolated
}
\description{
ms_synchronize_timestep both summarizes macrosheds to weekly, monthly, and 
annual time scales and interpolates macrosheds data to a finer time step such 
as sub daily
}
\details{
ms_synchronize_timestep is used to convert data into a finer
    temporal resolution, such as taking a weekly grab sample and interpolating 
    the values to have daily data, or to aggregate data into large intervals 
    such as monthly and annual sums or means. Macrosheds data undergoes temporal
    synchronization in our data processing pipeline. Our data is aggregated/imputed 
    to a daily time step for all variables and the impute limit for discharge and 
    precipitation is 3 days and for chemistry variables 15 days.
}
\author{
Spencer Rhea, \email{spencerrhea41@gmail.com}

Mike Vlah

Wes Slaughter
}
