% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_undo_scale_flux_by_area.R
\name{ms_undo_scale_flux_by_area}
\alias{ms_undo_scale_flux_by_area}
\title{Un-scale stream or precipitation chemical flux by watershed area}
\usage{
ms_undo_scale_flux_by_area(d)
}
\arguments{
\item{d}{\code{data.frame}. A \code{data.frame} in MacroSheds format (see \href{https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=1262}{MacroSheds EDI}
containing stream_flux_inst_scaled or precip_flux_inst_scaled data. A \code{data.frame} in
MacroSheds format is generated by \code{\link[=ms_load_product]{ms_load_product()}}.}
}
\value{
returns a \code{data.frame} in MacroSheds format with stream or precipitation chemical flux in kg/d.
}
\description{
MacroSheds chemical flux is scaled to watershed area by default (kg/ha/d).
This function converts chemical flux to kg/d. If You generated precipitation chemical
flux yourself with \code{\link[=ms_calc_watershed_precip]{ms_calc_watershed_precip()}}, it is in kd/d and this function is not needed.
}
\details{
MacroSheds reports all chemical flux on a per area basis (kg/ha/d) to
facilitate comparison of watersheds of different sizes. Sometimes un-scaled
fluxes are of interest. This function will convert the units of a native
MacroSheds chemical flux dataset from kg/ha/d to kg/d. To download
MacroSheds data, see \code{\link[=ms_download_core_data]{ms_download_core_data()}}.
}
\examples{
### Load some MacroSheds data:
ms_root = 'data/macrosheds'
ms_download_core_data(macrosheds_root = ms_root,
                      domains = 'hbef')
d <- ms_load_product(macrosheds_root = ms_root, 
                     prodname = 'stream_flux_inst_scaled', 
                     domains = 'hbef',
                     filter_vars = 'NO3_N')

d <- ms_undo_scale_flux_by_area(d)
}
\seealso{
\code{\link[=ms_scale_flux_by_area]{ms_scale_flux_by_area()}}
}
\author{
Spencer Rhea

Mike Vlah, \email{vlahm13@gmail.com}

Wes Slaughter
}
